Improve debuggability by linking the cause
While debugging a problem using the Generic JMS RA to connect AMQP to
Rabbit MQ the lack of visible linked exceptions in the stack made it
impossible to determine why things were not working. First the
createContext method threw away the caught exceptions completely, and
then the JMSSessionFactoryImpl set the linked exception, but this was
hidden when stack traces were output in logs, so by also calling
initCause all the useful failure exceptions suddenly became visible.