Move setting of client id to before session is created
Trying to use Apache Qpid wrapped in the RA results in an exception
because the client id is set after the session is created. The JMS spec
is ambiguous on setting client id after creating a connection, it
advises against it prefering a connection factory config instead, but
setting after the connection has done any reaal work is bad. QPID JMS
client seems to consider after a session is created to count, so moving
this into the createSession method just before the QPID session is
created.