Support Apache QPID JNDI paramaters
Apache QPID supports JNDI config where the connection factory name has a
value of a url with name value pairs. The syntax is here https://qpid.apache.org/releases/qpid-jms-0.25.0/docs/index.html

The parsing in the generic JMS assumes the value does not include an =
and if it does throws the entry away. This update changes to split name
from value based on the first = and allow equals later on. This allows
Apache QPID JNDI Parameters to work.